move env to a separate file
